+++ srcs/ft_lstnew.c
@@ -6,13 +6,13 @@
 /*   By: kfukuhar <kfukuhar@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/12 19:09:24 by kfukuhar          #+#    #+#             */
-/*   Updated: 2024/06/22 21:52:12 by kfukuhar         ###   ########.fr       */
+/*   Updated: 2024/06/22 21:55:40 by kfukuhar         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/push_swap.h"
 
-t_stack	*ft_lstnew(long int i)
+t_stack	*ft_lstnew(int i)
 {
 	t_stack	*node;
